feat(v0): POST /items, update OAS
Write a new JSON object file in the collection (using fs.writeFile),
with file name corresponding to `key` field provided in the request
body.

Includes a path traversal safeguard that works by normalizing the path
of the file to be written and checking that its dirname matches the
collection directory.