feat(utils): implement metadata route
The metadata route returns an informative summary of metadata fetched
and scraped from the url. The url is specified using a required query
parameter. A subset of fields can be requested using an optional query
parameter.

The backend script uses BetaHuhn/metadata-scraper and a set of rules to
shape the returned data.