refactor(v0): decouple Swagger UI from OAS generation (jsdoc)
Add a route for the OpenAPI Specification document to the root of the
API version (/v0/oas3.json). The spec is still generated by
swagger-jsdoc.

Load the OAS document to the Swagger UI (/v0/api-docs) from URL. See
"Load swagger from url" and "Link to swagger document" in
swagger-ui-express documentation.