feat(v1): add knex, sqlite connection, v1 route
Export a knex connection (knex.js) to sqlite (knexfile.js), to be
imported by v1 router and used for queries and updates on the database.
Migrations and seeds are handled in the data-tier repo.

Created a v1 route on the server (server.js, routes/index.js), pointing
to a new router (routes/index.js).

Implemented OpenAPI document generation for v1 (routes/v1/index.js,
api-definition.json), created a route to the document, linked Swagger UI
to the document (api-docs.js)